refactor: 소수도 추출할 수 있도록 변경 및 커스텀 구분자는 소수점을 쓸 수 없도록 제한

+++ src/main/java/calculator/view/OutputView.java
@@ -1,7 +1,9 @@
 package calculator.view;
 
+import java.math.BigDecimal;
+
 public class OutputView {
-    public void printResult(int total) {
+    public void printResult(BigDecimal total) {
         System.out.println("결과 : " + total);
     }
 }
